Backend/requirements.txt
asgiref==3.8.1
Django==5.1.7
sqlparse==0.5.3
typing_extensions==4.13.0
djangorestframework==3.14.0
Pillow==10.1.0  
django-cors-headers==4.2.0
gunicorn==21.2.0  
python-dotenv==1.0.0  
